+++ easypeel-method-security/src/main/java/org/easypeelsecurity/core/MethodBan.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.epsec.core;
+package org.easypeelsecurity.core;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
