+++ easypeel-method-security/src/main/java/org/easypeelsecurity/engine/MethodBanManager.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.epsec.engine;
+package org.easypeelsecurity.engine;
 
 import java.time.LocalDateTime;
 import java.util.HashMap;
