+++ easypeel-method-security/src/main/resources/META-INF/services/javax.annotation.processing.Processor
@@ -1,1 +1,1 @@
-org.epsec.core.MethodBanProcessor
+org.easypeelsecurity.core.MethodBanProcessor
