+++ easypeel-method-security/src/test/java/org/easypeelsecurity/core/MethodBanProcessorTest.java
@@ -14,13 +14,22 @@
  * limitations under the License.
  */
 
-package org.epsec.core;
+package org.easypeelsecurity.core;
 
 import static com.google.testing.compile.CompilationSubject.assertThat;
 import static com.google.testing.compile.Compiler.javac;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.doThrow;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
 
+import java.io.IOException;
 import java.util.stream.Stream;
 
+import javax.annotation.processing.Filer;
+import javax.annotation.processing.Messager;
+import javax.tools.Diagnostic;
 import javax.tools.JavaFileObject;
 
 import org.junit.jupiter.api.Test;
@@ -30,6 +39,7 @@ import org.junit.jupiter.params.provider.MethodSource;
 
 import com.google.testing.compile.Compilation;
 import com.google.testing.compile.JavaFileObjects;
+import com.squareup.javapoet.JavaFile;
 
 @SuppressWarnings("checkstyle:FinalLocalVariable")
 class MethodBanProcessorTest {
@@ -42,7 +52,7 @@ class MethodBanProcessorTest {
 
     // when
     Compilation compilation = javac()
-        .withProcessors(new MethodBanProcessor())
+        .withProcessors(new org.easypeelsecurity.core.MethodBanProcessor())
         .withOptions("-source", javaVersion, "-target", javaVersion)
         .compile(src);
 
@@ -51,10 +61,26 @@ class MethodBanProcessorTest {
   }
 
   @Test
+  void writeToOccurIoException() throws IOException {
+    // given
+    final Messager messager = mock(Messager.class);
+    final JavaFile javaFile = mock(JavaFile.class);
+    doThrow(IOException.class).when(javaFile).writeTo(any(Filer.class));
+
+    // when & then
+    JavaFileObject src = JavaFileObjects.forResource("before/MethodBanHappy.java");
+    Compilation compilation = javac()
+        .withProcessors(new org.easypeelsecurity.core.MethodBanProcessor())
+        .compile(src);
+
+    verify(messager, times(1)).printMessage(Diagnostic.Kind.ERROR, any());
+  }
+
+  @Test
   void compileHappy() {
     JavaFileObject src = JavaFileObjects.forResource("before/MethodBanHappy.java");
     Compilation compilation = javac()
-        .withProcessors(new MethodBanProcessor())
+        .withProcessors(new org.easypeelsecurity.core.MethodBanProcessor())
         .compile(src);
     assertThat(compilation).succeededWithoutWarnings();
   }
@@ -64,7 +90,7 @@ class MethodBanProcessorTest {
   void testMethodBanProcessorHappy(String fileName) {
     JavaFileObject src = JavaFileObjects.forResource("before/" + fileName);
     Compilation compilation = javac()
-        .withProcessors(new MethodBanProcessor())
+        .withProcessors(new org.easypeelsecurity.core.MethodBanProcessor())
         .compile(src);
     assertThat(compilation).succeededWithoutWarnings();
   }
@@ -81,7 +107,7 @@ class MethodBanProcessorTest {
   void testMethodBanProcessor(String fileName, String expectedErrorMessage) {
     JavaFileObject src = JavaFileObjects.forResource("before/" + fileName);
     Compilation compilation = javac()
-        .withProcessors(new MethodBanProcessor())
+        .withProcessors(new org.easypeelsecurity.core.MethodBanProcessor())
         .compile(src);
     assertThat(compilation).hadErrorContaining(expectedErrorMessage);
   }
