+++ easypeel-method-security/src/test/java/org/easypeelsecurity/engine/FqcnTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.epsec.engine;
+package org.easypeelsecurity.engine;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
