+++ easypeel-method-security/src/test/resources/before/MethodBanHappy.java
@@ -1,6 +1,6 @@
 package com.example;
 
-import org.epsec.core.MethodBan;
+import org.easypeelsecurity.core.MethodBan;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
