+++ gradle.properties
@@ -1,5 +1,5 @@
-easypeelMethodSecurityVersion=0.0.3
-rootPackageName=io.github.easypeel-security
+easypeelMethodSecurityVersion=0.0.4
+rootPackageName=org.easypeelsecurity
 easypeelMethodSecurity=easypeel-method-security
 
 junitVersion=5.10.1
